+++ .github/workflows/publish-npm-packages.yml
@@ -1,0 +1,55 @@
+name: Publish NPM Packages
+on:
+  workflow_dispatch:
+    inputs:
+      ampversion:
+        description: 'AMP version'
+        required: true
+      tag:
+        description: 'NPM package tag (latest | nightly)'
+        required: true
+jobs:
+  publish:
+    environment: NPM_TOKEN #TODO: change environment name in repo settings
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        extension:
+          [
+            'amp-selector',
+            'amp-accordion',
+            'amp-date-countdown',
+            'amp-date-display',
+            'amp-fit-text',
+            'amp-inline-gallery',
+            'amp-instagram',
+            'amp-lightbox',
+            'amp-timeago',
+            'amp-youtube',
+          ]
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          ref: ${{ github.events.inputs.ampversion }}
+      - uses: actions/setup-node@v2
+        with:
+          node-version: 14.x
+          registry-url: https://registry.npmjs.org
+      - name: Build package
+        run: |
+          node ${{ github.workspace }}/build-system/npm-publish/build-npm-binaries.js ${{ matrix.extension }}
+          node ${{ github.workspace }}/build-system/npm-publish/write-package-json.js ${{ matrix.extension }} ${{ github.event.inputs.ampversion }}
+      - name: Publish v1
+        run: |
+          cd ${{ github.workspace }}/extensions/${{ matrix.extension }}/1.0 
+          npm publish --access public --tag ${{ github.event.inputs.tag }}
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
+      - name: Publish v2 if it exists
+        run: |
+          if $(test -d ${{ github.workspace }}/extensions/${{ matrix.extension }}/2.0); then      
+            cd ${{ github.workspace }}/extensions/${{ matrix.extension }}/2.0
+            npm publish --access public --tag ${{ github.event.inputs.tag }}
+          fi
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
