+++ build-system/npm-publish/write-package-jsons.js
@@ -21,7 +21,7 @@
  */
 
 const [extension, ampVersion] = process.argv.slice(2);
-const {log} = require('../../../build-system/common/logging');
+const {log} = require('../common/logging');
 const {stat, writeFile} = require('fs/promises');
 const {valid} = require('semver');
 
@@ -38,7 +38,11 @@ async function writePackageJson(extensionVersion) {
   const minor = ampVersion.slice(0, 10);
   const patch = Number(ampVersion.slice(-3)); // npm trims leading zeroes in patch number, so mimic this in package.json
   const version = `${major}.${minor}.${patch}`;
-  if (!valid(version) || ampVersion.length != 13 || extensionVersionArr[1] !== '0') {
+  if (
+    !valid(version) ||
+    ampVersion.length != 13 ||
+    extensionVersionArr[1] !== '0'
+  ) {
     log(
       'Invalid semver version',
       version,
